requirements.txt
Django
django-ipware
django-ratelimit
celery
redis
geoip2
scikit-learn
requests
